Refactor and add admin routes

+++ Services/Users/AdminRepository/AdminRepositoryService.cs
@@ -32,7 +32,8 @@ namespace RealEstate.Services.Users.AdminRepository
             var users = await _userManager.Users.ToListAsync().ConfigureAwait(false);
             foreach (var user in users)
             {
-                await _userManager.DeleteAsync(user).ConfigureAwait(false);
+                if (await _userManager.IsInRoleAsync(user, Roles.User).ConfigureAwait(false))
+                    await _userManager.DeleteAsync(user).ConfigureAwait(false);
             }
             var environmentPath = _imageService.GetLocalImagesFullPath("auth");
             if (Directory.Exists(environmentPath))
@@ -45,13 +46,22 @@ namespace RealEstate.Services.Users.AdminRepository
             }
         }
 
-        public async Task<IdentityResult> AssignRole(User user, IEnumerable<User> users)
+        public async Task<IdentityResult> AssignRole(User user)
         {
+            var allUsers = await GetUsersAsync().ConfigureAwait(false);
             bool userFlagged = false;
-            if (users.Count() == 1)
+            if (allUsers.Count() == 1)
                 userFlagged = true;
             return await _userManager.AddToRoleAsync(user, userFlagged ? Roles.Admin : Roles.User).ConfigureAwait(false);
         }
+
+        public async Task<IdentityResult> PromoteUser(User user) {
+            var result = await _userManager.AddToRoleAsync(user, Roles.Admin).ConfigureAwait(false);
+            if (!result.Succeeded)
+                return IdentityResult.Failed();
+            return await _userManager.RemoveFromRoleAsync(user,Roles.User).ConfigureAwait(false);
+        }
+        
 
         public async Task<List<UserProfileImage>> GetUserProfileImageListAsync() =>
       await _userIdentityContext
@@ -60,6 +70,8 @@ namespace RealEstate.Services.Users.AdminRepository
      .OrderByDescending(userProfileImg => userProfileImg.Id)
      .ToListAsync()
      .ConfigureAwait(false);
+
+        public async Task<bool> IsAdmin(User user) => await _userManager.IsInRoleAsync(user, Roles.Admin).ConfigureAwait(false);
 
 
     }
